bitcoind: move helper functions to a new 'utils' submodule
The poller submodule is already kinda hairy, and having the helpers
didn't help readability.
